Add EntityAssets struct to common models and update fitness models to use it
- Introduced the EntityAssets struct to represent uploaded assets for an entity, including fields for images and videos.
- Updated fitness models to import and utilize the new EntityAssets struct, enhancing the organization and clarity of asset management in the codebase.
